Rename to train_sft_native.py
